Experimenting with making a Makfile
Made a Makefile in not-package to compile up the analyses and the document, but in the end felt that the inability to easily define multiple targets that are created with a single rule was a serious drawback.

I think it is better to just stick with R and write a script to compile everything.  So I will probably delete the Makefile in the next commit.