Simulation of Mixture Model with Noncontaminated Data
Created a Rscript, sim_mixture, that randomly picks individuals from the baseline and then tests the mixture MCMC.  The code is explained a README in the notes directory.  After running a few test, it seems that the model is not too great at assigning individuals to the right RepUnit (~50% accuracy).