Moved make_mixture to not_package and started simulation.
I moved make_mixture to not_package/simulations to a script called mixed_simulation_functions (I am putting all the simulation functions there).
Started working on the simulation.  Made code to function to get individuals using the ca_fishery_proportions and made code to make means and things out of the MCMC data.  Started working on actually running the simulations and making data frames, but haven't had a chance to check if it is working properly.